requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sentence-transformers
numpy
faiss-cpu
PyPDF2
pdfplumber
python-docx
python-magic
beautifulsoup4
httpx
groq
chromadb
psutil